+++ probes/entries.go
@@ -30,7 +30,7 @@ import (
 	"github.com/ossf/scorecard/v4/probes/fuzzedWithPythonAtheris"
 	"github.com/ossf/scorecard/v4/probes/fuzzedWithRustCargofuzz"
 	"github.com/ossf/scorecard/v4/probes/fuzzedWithSwiftLibFuzzer"
-	"github.com/ossf/scorecard/v4/probes/packagedNpmWithGitHubWorkflow"
+	"github.com/ossf/scorecard/v4/probes/packagedWithAutomatedWorkflow"
 	"github.com/ossf/scorecard/v4/probes/securityPolicyContainsLinks"
 	"github.com/ossf/scorecard/v4/probes/securityPolicyContainsText"
 	"github.com/ossf/scorecard/v4/probes/securityPolicyContainsVulnerabilityDisclosure"
@@ -79,7 +79,7 @@ var (
 		fuzzedWithPropertyBasedJavascript.Run,
 	}
 	Packaging = []ProbeImpl{
-		packagedNpmWithGitHubWorkflow.Run,
+		packagedWithAutomatedWorkflow.Run,
 	}
 )
